[FLINK-13632] Port EnumValueSerializer upgrade test to TypeSerializerUpgradeTestBase
